+++ src/views/@common/hooks/useInterceptor.ts
@@ -1,11 +1,11 @@
-import { isAxiosError } from 'axios';
 import { useEffect } from 'react';
 
-import removeToken from '../utils/removeToken';
-
-import api, { getRefreshToken, getToken, setRefreshToken, setToken } from './api';
+import api, { getToken } from './api';
+import usePostRefresh from './usePostRefresh';
 
 const useInterceptor = () => {
+  const postRefresh = usePostRefresh();
+
   useEffect(() => {
     api.interceptors.request.use((config) => {
       const accessToken = getToken();
@@ -14,27 +14,6 @@ const useInterceptor = () => {
       }
       return config;
     });
-
-    const postRefresh = async () => {
-      const accessToken = getToken();
-      const refreshToken = getRefreshToken();
-
-      try {
-        const response = await api.post('/auth/refresh', {
-          accessToken: accessToken,
-          refreshToken: refreshToken,
-        });
-        const newToken = response.data.data;
-        setToken(newToken.accessToken);
-        setRefreshToken(newToken.refreshToken);
-      } catch (err) {
-        // refresh api에서 발생한 에러 처리
-        if (isAxiosError(err) && err.response?.status === 401) {
-          console.log(err);
-          removeToken();
-        }
-      }
-    };
 
     api.interceptors.response.use(
       (res) => res,
